Refactor: Move page templates to templates directory
